finally first render

+++ vec_comp.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-int		product(t_vec3 a, t_vec3 b)
+double	product(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
